Finished logout and login endpoints and added more @types

+++ src/routes/Session.ts
@@ -1,18 +1,19 @@
 // eslint-disable-next-line no-unused-vars
-import express, { Response } from 'express';
+import express, { Response, Request } from 'express';
 import bcrypt from 'bcrypt';
 import nanoid from 'nanoid';
 import StrongParams from '../middleware/StrongParam';
-import SessionMiddleware from '../middleware/AuthSession';
 import StrongParam from '../middleware/StrongParam';
 import User from '../models/User';
 import Session from '../models/Session';
 const router = express.Router();
 
-router.post('/logout', SessionMiddleware, async (_, res: Response) => {
-  const sid = res.locals.sessionCookie.sid;
-  await Session.findOneAndDelete({ sid, });
-  return res.status(200).send('Session Logout');
+router.post('/logout', async (req: Request, res: Response) => {
+  const sid = req.signedCookies?.session;
+  if (sid) {
+    await Session.findOneAndDelete({ sid, });
+  }
+  return res.clearCookie('session').sendStatus(200);
 });
 
 const loginStrongParams = {
@@ -24,7 +25,8 @@ router.post('/login', StrongParams(loginStrongParams), async (_, res: Response) 
     const { email, password } = res.locals.body;
     const user = await User.findOne({ email });
     if (await bcrypt.compare(password, user.passwordDigest)) {
-      return res.sendStatus(200);
+      const sid = await createSession(user.uid);
+      return res.cookie('session', sid, { maxAge: 86400000, signed: true, httpOnly: true }).sendStatus(200);
     } else {
       return res.sendStatus(401);
     }
@@ -72,12 +74,8 @@ router.post('/signup', StrongParam(signUpStrongParams), async (_, res: Response)
       isTechnician
     }).save();
     if (newUser) {
-      let sid = '';
-      do {
-        sid = nanoid();
-      } while (await Session.findOne({ sid, }));
-      await new Session({ uid: newUser.uid, sid }).save();
-      return res.status(200).cookie('session', sid, { maxAge: 86400000 }).send(newUser);
+      const sid = await createSession(newUser.uid);
+      return res.cookie('session', sid, { maxAge: 86400000, signed: true, httpOnly: true }).sendStatus(200);
     } else {
       throw new Error('Unable to create new User');
     }
@@ -86,5 +84,14 @@ router.post('/signup', StrongParam(signUpStrongParams), async (_, res: Response)
     return res.sendStatus(500);
   }
 });
+
+async function createSession(uid: string): Promise<string> {
+  let sid;
+  do {
+    sid = nanoid();
+  } while (await Session.findOne({ sid, }));
+  await new Session({ uid, sid }).save();
+  return sid;
+}
 
 export default router;
